// apps/web/src/lib/api-schema.d.ts
export interface components {
    schemas: {
// ...
        SourceTranscript: {
            /** Duration Ms */
            duration_ms: number;
            /** Segments */
            segments: components["schemas"]["TranscriptSegment"][];
            /**
             * Schema Version
             * @default 0.1.0
             * @constant
             */
            schema_version: "0.1.0";
            /** Source Ids */
            source_ids: string[];
        };
// ...
        TranscriptSegment: {
            /** Text */
            text: string;
            /** Clip Start Ms */
            clip_start_ms: number;
            /** Clip End Ms */
            clip_end_ms: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}